readme: Use markdown syntax instead of <pre>
The <pre> tag is stripped by Gitiles causing content to be missing.
It also isn't well-supported on Doxygen which just renders it as
plain text instead of highlighted.

The only reason I used <pre> here was to satisfy Phabricator
with highlighting optimised for GitHub, neither of which are
canonical for us.

The main thing that matters is that it renders fine in Doxygen
and on Gitiles.

Change-Id: Iee9c320067c2d1364c440e425067e331eebeb4e9